refactor: Extract dashboard table columns
