+++ src/app/dashboard/users/page.tsx
@@ -1,7 +1,9 @@
 import type { Metadata } from "next";
 import { headers } from "next/headers";
 import { notFound } from "next/navigation";
+import { Users as UsersIcon } from "lucide-react";
 import { getTranslations } from "next-intl/server";
+import DashboardPageIntro from "@/components/dashboard/PageIntro";
 import DashboardSiteHeader from "@/components/dashboard/SiteHeader";
 import DashboardUsersManager from "@/components/dashboard/UsersManager";
 import { getCurrentUserFromHeaders } from "@/lib/server/auth-session";
@@ -39,7 +41,13 @@ export default async function DashboardUsersPage() {
         parent={{ label: tCommon("labels.dashboard"), href: "/dashboard" }}
         title={t("pages.users")}
       />
-      <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
+      <div className="flex flex-1 flex-col gap-6 p-4 pt-0">
+        <DashboardPageIntro
+          icon={UsersIcon}
+          title={t("pages.users")}
+          description={t("pages.usersIntro")}
+          accent="bg-emerald-500/10 text-emerald-600 dark:text-emerald-400"
+        />
         <DashboardUsersManager
           currentUserId={currentUser.id}
           initialUsers={users}
