+++ src/components/dashboard/ApiKeysManager.tsx
@@ -3,17 +3,27 @@
 import { useState } from "react";
 import { useTranslations } from "next-intl";
 import {
-  type ColumnDef,
   getCoreRowModel,
   getFilteredRowModel,
   getSortedRowModel,
   type SortingState,
   useReactTable,
 } from "@tanstack/react-table";
-import { ArrowUpDown, KeyRound } from "lucide-react";
+import { KeyRound } from "lucide-react";
+import {
+  formatDate,
+  formatDateTime,
+  getApiKeyStatus,
+  getApiKeysColumns,
+  getKeyLabel,
+  getOwnerLabel,
+  getStatusLabel,
+  getStatusVariant,
+  type ApiKeyStatus,
+  type Translate,
+} from "@/app/dashboard/api-keys/columns";
 import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import {
   Sheet,
   SheetContent,
@@ -23,36 +33,10 @@ import {
 } from "@/components/ui/sheet";
 import DataTable from "@/components/data-table/DataTable";
 import DataTableFacetFilter from "@/components/data-table/DataTableFacetFilter";
-import { dataTableSortButtonClassName } from "@/components/data-table/DataTableLayout";
 import DataTableToolbar from "@/components/data-table/DataTableToolbar";
 import type { AdminApiKey } from "@/lib/server/api-keys";
 
-type ApiKeyStatus = "active" | "expired" | "disabled";
-
 const statusFilterValues: ApiKeyStatus[] = ["active", "expired", "disabled"];
-
-function getApiKeyStatus(key: AdminApiKey): ApiKeyStatus {
-  if (!key.enabled) return "disabled";
-  if (key.expiresAt && new Date(key.expiresAt).getTime() <= Date.now()) {
-    return "expired";
-  }
-  return "active";
-}
-
-function getStatusVariant(
-  status: ApiKeyStatus
-): "default" | "muted" | "outline" {
-  if (status === "active") return "outline";
-  return "muted";
-}
-
-function getStatusLabel(status: ApiKeyStatus, t: (key: string) => string) {
-  return t(`statusValues.${status}`);
-}
-
-function getOwnerLabel(key: AdminApiKey) {
-  return key.ownerName?.trim() || key.ownerEmail?.trim() || key.ownerUserId;
-}
 
 function getOwnerInitials(key: AdminApiKey) {
   const name = key.ownerName?.trim();
@@ -63,33 +47,6 @@ function getOwnerInitials(key: AdminApiKey) {
     .join("")
     .slice(0, 2)
     .toUpperCase();
-}
-
-function getKeyLabel(key: AdminApiKey) {
-  return key.name ?? key.start ?? key.id;
-}
-
-function formatDate(value: string | null) {
-  if (!value) return "—";
-  try {
-    return new Intl.DateTimeFormat("en-GB", { dateStyle: "medium" }).format(
-      new Date(value)
-    );
-  } catch {
-    return value;
-  }
-}
-
-function formatDateTime(value: string | null) {
-  if (!value) return "—";
-  try {
-    return new Intl.DateTimeFormat("en-GB", {
-      dateStyle: "medium",
-      timeStyle: "short",
-    }).format(new Date(value));
-  } catch {
-    return value;
-  }
 }
 
 function formatRateLimitWindow(
@@ -122,116 +79,7 @@ export default function DashboardApiKeysManager({
   const [selectedStatuses, setSelectedStatuses] = useState<ApiKeyStatus[]>([]);
   const [inspectKey, setInspectKey] = useState<AdminApiKey | null>(null);
 
-  const columns: ColumnDef<AdminApiKey>[] = [
-    {
-      id: "key",
-      accessorFn: (row) => getKeyLabel(row),
-      meta: { className: "w-[28%] min-w-48" },
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          size="sm"
-          className={dataTableSortButtonClassName}
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-        >
-          {t("table.key")}
-          <ArrowUpDown className="text-muted-foreground ml-1 size-3.5" />
-        </Button>
-      ),
-      cell: ({ row }) => (
-        <div className="min-w-0">
-          <p className="truncate text-sm font-medium">
-            {getKeyLabel(row.original)}
-          </p>
-          <p className="text-muted-foreground truncate font-mono text-xs">
-            {row.original.prefix ?? ""}
-            {row.original.start ?? ""}…
-          </p>
-        </div>
-      ),
-    },
-    {
-      id: "owner",
-      accessorFn: (row) => getOwnerLabel(row),
-      header: t("table.owner"),
-      meta: { className: "w-[25%] min-w-44" },
-      cell: ({ row }) => (
-        <div className="min-w-0">
-          <p className="truncate text-sm">{getOwnerLabel(row.original)}</p>
-          <p className="text-muted-foreground truncate text-xs">
-            {row.original.ownerEmail ?? row.original.ownerUserId}
-          </p>
-        </div>
-      ),
-    },
-    {
-      id: "status",
-      accessorFn: (row) => getApiKeyStatus(row),
-      header: t("table.status"),
-      meta: { className: "w-28" },
-      cell: ({ row }) => {
-        const status = getApiKeyStatus(row.original);
-        return (
-          <Badge variant={getStatusVariant(status)}>
-            {getStatusLabel(status, t)}
-          </Badge>
-        );
-      },
-    },
-    {
-      id: "requests",
-      accessorKey: "requestCount",
-      meta: { className: "w-28" },
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          size="sm"
-          className={dataTableSortButtonClassName}
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-        >
-          {t("table.requests")}
-          <ArrowUpDown className="text-muted-foreground ml-1 size-3.5" />
-        </Button>
-      ),
-      cell: ({ row }) => (
-        <span className="text-sm tabular-nums">
-          {row.original.requestCount.toLocaleString()}
-        </span>
-      ),
-    },
-    {
-      id: "lastUsed",
-      accessorKey: "lastRequest",
-      meta: { className: "w-36" },
-      header: ({ column }) => (
-        <Button
-          variant="ghost"
-          size="sm"
-          className={dataTableSortButtonClassName}
-          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
-        >
-          {t("table.lastUsed")}
-          <ArrowUpDown className="text-muted-foreground ml-1 size-3.5" />
-        </Button>
-      ),
-      cell: ({ row }) => (
-        <span className="text-muted-foreground text-xs">
-          {formatDateTime(row.original.lastRequest)}
-        </span>
-      ),
-    },
-    {
-      id: "expires",
-      accessorKey: "expiresAt",
-      meta: { className: "w-32 hidden lg:table-cell" },
-      header: t("table.expires"),
-      cell: ({ row }) => (
-        <span className="text-muted-foreground text-xs">
-          {formatDate(row.original.expiresAt)}
-        </span>
-      ),
-    },
-  ];
+  const columns = getApiKeysColumns({ t: t as unknown as Translate });
 
   // eslint-disable-next-line react-hooks/incompatible-library
   const table = useReactTable({
@@ -373,7 +221,10 @@ export default function DashboardApiKeysManager({
                       label: t("panel.fields.status"),
                       value: (
                         <Badge variant={getStatusVariant(inspectStatus)}>
-                          {getStatusLabel(inspectStatus, t)}
+                          {getStatusLabel(
+                            inspectStatus,
+                            t as unknown as Translate
+                          )}
                         </Badge>
                       ),
                     },
